+++ CHANGELOG.md
@@ -32,6 +32,8 @@ changes.
 
 - Enable blockfrost integration for hydra-node.
 
+- Fix head status in post abort greetings output.
+
 ## [0.21.0] - 2025-04-28
 
 - New metric for counting the number of active peers: `hydra_head_peers_connected`